Rename measure2/ folder to measure/
